refactor(bcd): phase M2 — extract WorkspaceServices factory
Moves the ~300-line init block out of internal/cmd/serve.go into
server/build_services.go:BuildWorkspaceServices and a sibling
server/stats_collector.go with the docker/tmux/token metric
collectors. RunServer is now a thin orchestrator: build Globals,
call the factory, wire handlers, start.

server.Globals captures the truly workspace-agnostic dependencies
(Registry, Stats, Deps, GlobalHub, Build) that the factory needs
but that persist for bcd's whole lifetime.

Behaviour is unchanged; single-workspace paths still serve the
launch workspace's services. Phase M3 will route requests through
context and phase M5 will let the factory run for any registered
workspace.
